Loose cleanups from player reference

+++ Source/player.h
@@ -77,7 +77,7 @@ enum inv_body_loc : uint8_t {
 	NUM_INVLOC,
 };
 
-enum class player_graphic : uint16_t {
+enum class player_graphic : uint8_t {
 	Stand,
 	Walk,
 	Attack,
